renamed index.js and test.js and created EvenOrOdd problem
