+++ EvenOrOdd.js
@@ -1,0 +1,6 @@
+export function evenOrOdd(number) {
+  if (Math.floor(number/2) !== 0){
+    return "Odd";
+  }
+  return "Even";
+}
