+++ test/EvenOrOddTest.js
@@ -1,0 +1,19 @@
+import { evenOrOdd } from "../EvenOrOdd.js";
+
+it('Even number test (2)', function() {
+    const result = evenOrOdd(2);
+    const expectedResult = "Even";
+    equal(result, expectedResult);
+});
+
+it('Even number test (0)', function() {
+    const result = evenOrOdd(2);
+    const expectedResult = "Even";
+    equal(result, expectedResult);
+});
+
+it('Even number test (0)', function() {
+    const result = evenOrOdd(2);
+    const expectedResult = "Even";
+    equal(result, expectedResult);
+});
